Load and save the fpu registers (for copies to/from userspace) using   helper functions in arch/x86/x86/fpu.c They (hopefully) ensure that we write to the entire buffer and don't load   values that might cause faults in kernel. Also zero out the 'pad' field of the i386 mcontext fp area that I think   once contained the registers of any Weitek fpu.   Dunno why it wasn't pasrt of the union. Some of these copies could be removed if the code directly copied the save   area to/from userspace addresses.

+++ sys/arch/i386/include/mcontext.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: mcontext.h,v 1.11 2014/02/04 22:21:35 dsl Exp $	*/
+/*	$NetBSD: mcontext.h,v 1.12 2014/02/15 22:20:42 dsl Exp $	*/
 
 /*-
  * Copyright (c) 1999 The NetBSD Foundation, Inc.
@@ -86,7 +86,7 @@ typedef struct {
 		} __fp_xmm_state;	/* x87 and xmm regs in fxsave format */
 		int	__fp_fpregs[128];
 	} __fp_reg_set;
-	long	__fp_pad[33];			/* Historic padding */
+	int 	__fp_pad[33];			/* Historic padding */
 } __fpregset_t;
 __CTASSERT(sizeof (__fpregset_t) == 512 + 33 * 4);
